fix: fixed issue with sign function refactoring

+++ contracts/libraries/Account.sol
@@ -315,7 +315,7 @@ library Account {
     }
 
     function sign(int256 value) internal pure returns (int256) {
-        return value > 0 ? 1 : -1;
+        return value > 0 ? int256(1) : int256(-1);
     }
 
     function liquidateTokenPosition(
